+++ app/errors/index.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var app = module.exports = express();
-var config = require('../../config');
+var config = require('app/config');
 
 app.set('views', config.views);
 app.set('view engine', 'jade');
