+++ app/home/index.js
@@ -1,11 +1,11 @@
 var express = require('express');
 var app = module.exports = express();
-var config = require('../../config');
+var config = require('app/config');
 
 app.set('views', config.views);
 app.set('view engine', 'jade');
 
 app.route('/')
-    .get(function(req, res) {
-      res.render('home/index')
-    });
+  .get(function(req, res) {
+    res.render('home/index')
+  });
