+++ app/shared/layout.jade
@@ -17,7 +17,6 @@ html(lang="en")
     title #{globalVars.title}
 
   body
-    //- include navbar
 
     if (flash)
       p #{flash}
