+++ app/users/middle.js
@@ -1,7 +1,7 @@
-var Blog = require('app/blog/model');
+var User = require('app/users/model');
 
 function findAll(req, res, next) {
-  Blog.find({}, null, {sort: {created_date: -1}}, function(err, data) {
+  User.find({}, null, {sort: {created_date: -1}}, function(err, data) {
     if (err) return next(err);
     res.locals.blog = data;
     return next();
@@ -10,7 +10,7 @@ function findAll(req, res, next) {
 
 function find(req, res, next) {
   var query = (req.params.id ? {_id: req.params.id} : {});
-  Blog.findOne(query, null, {sort: {created_date: -1}}, function(err, data) {
+  User.findOne(query, null, {sort: {created_date: -1}}, function(err, data) {
     if (err) return next(err);
     res.locals.blog = data;
     return next();
